+++ SURFSharekit.Net/Models/RepoItem/SURFSharekitAttributes.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace SURFSharekit.Net.Models;
+namespace SURFSharekit.Net.Models.RepoItem;
 
 public class SURFSharekitAttributes
 {
