+++ SURFSharekit.Net/Models/RepoItem/SURFSharekitEducationalLevel.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace SURFSharekit.Net.Models;
+namespace SURFSharekit.Net.Models.RepoItem;
 
 public class SURFSharekitEducationalLevel
 {
