+++ SURFSharekit.Net/Models/RepoItem/SURFSharekitFile.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace SURFSharekit.Net.Models;
+namespace SURFSharekit.Net.Models.RepoItem;
 
 public class SURFSharekitFile
 {
