+++ SURFSharekit.Net/Models/RepoItem/SURFSharekitRepoItemLinks.cs
@@ -5,7 +5,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace SURFSharekit.Net.Models;
+namespace SURFSharekit.Net.Models.RepoItem;
 
 public class SURFSharekitRepoItemLinks
 {
