+++ SURFSharekit.Net/Models/WebhooksByDocumentation/SURFSharekitWebhookPayloadType.cs
@@ -5,7 +5,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace SURFSharekit.Net.Models;
+namespace SURFSharekit.Net.Models.WebhooksByDocumentation;
 
 [JsonConverter(typeof(JsonStringEnumConverter<SURFSharekitWebhookPayloadType>))]
 public enum SURFSharekitWebhookPayloadType
